Fix module namespace issues
In the previous way defining the Railtie classes:

    class MiqPerformance::FooBarRailtie < Railtie

The code, while more terse, essentially does this:

    MiqPerformance::FooBarRailtie = Class.new

Meaning, it expects that `MiqPerformance` is defined, and then defines a
class within the namespace of that module.  When compared to the
traditional namespace style:

    module MiqPerformance
      class FooBarRailtie < Railtie
    end

Which will define the `MiqPerformance` module (if it doesn't already
exist) or "MonkeyPatch" it to define the class (constant) within the
context of that module.  This is the more moduluar approach as we don't
know if `lib/miq_performance.rb` will be loaded ahead of time or not,
which would normally define `MiqPerformance`.  This style removes the
need for that.