Reset config after all tests
Most tests will be testing different configs and variations, so reseting
and reloading the config after each test, requiring it to be reloaded
for each test, is something that is convenient enough (and not a huge
performance hit on the suite overall) to do every time.

Note:  The performance will not be realized in the `after` filter, but in
the subsequent test that now has to reload the config.  Again, this is
just a file read operation, so it should be an acceptable performance
penalty, if slightly obscure to a new dev.