Add stack trace info to active_record_queries info
When configured.  Currently disabled by default, since this is a more of
a developer feature versus a feature needed in production.

Trace currently removes anything that isn't in the `Rails.root`, and
removes file path info that includes the `Rails.root` from the saved
trace to save on storage space.