Adds skip_schema_queries to the config
Defaults it to false.

Also fixes a bug with MiqPerformance::Configuration#[] where if a config
value was falsey, it would return the default value for the config
instead of the user configured value.  This was because the code was
doing a simple `||`, and switching it to a `.fetch` fixes that bug.