Handle commands that don't load in help sub cmd
There will some commands that are provided by miqperf that require
optional dependant gems, so allow `miqperf` to function in both
situations:  with and without the dependency.

Instead of displaying help docs for them regardless if sub command will
actually work, try to load them, but if a NameError is raised (the
command files themselves should handle not defining the class if the
dependency doesn't exists), then rescue and continue on.