Fix ~/.miq_performance issue in tests
There was a non-deterministic spec failure when the specs were run as a
suite.  This was caused by certain specs making use of the
configuration, assuming the default, but not stubbing out the
config_file_location and it was processed when running that test,
memoized, and then saved for future tests.

Also, it makes more sense to make sure this is cleared before each test
than waiting until a test is finished to have it clean up after itself.

Of note:  `if defined?` is used because not all specs load the
configuration class, so this constant might not be initialized when
running individual tests, or tests in a specific order.