Use MiqPerformance::Configuration[] for initialize
After setting the config, we want to make use of the `[]` instance
method since it does similar functionality to what was being done here,
but also makes use of "fetch", which avoids issues with the key existing
in the config, but being `falsey`.

For these keys, this might cause issues since they should never be
`nil`/`false`, but it will provide a more expected behavior to the
user/developer when these keys are that instead of jumping to the
DEFAULTS even though the key exists.