Adds StacktraceCleaners::Rails
Adds a Rails cleaner that uses Rails::BacktraceCleaner to filter
results.  This is expected to perform slower than the Simple cleaner,
but include a lot of sensible defaults that are nice when using Rails.

This also moves the stubbing of Rails.root to the spec helper to avoid
Ruby/rspec warnings.