Make ActiveSupportTimers report like the rest
Previously, the active_support_timers_save method was called during the
middle of ActiveSupport::Notification call, which is different from the
rest of the middleware which wrote it's collected data during the
`finish` call for that specific middleware.

To allow for a unified way of collecting the data, the
ActiveSupportTimers middleware has been updated to write to a temporary
Thread variable to store it's data until it can be used to write to the
Middleware Storage in it's finish method.