Adds MiddlewareStorage::Log
This is a simple (hopefully thread safe) logger that will aggregate
specific statistics and combine them and write to a single log file line
per request.  Meant for simple statistics gathering for all requests
that is machine parse-able, with the goal of limiting what data gets
written to make it relatively human parse-able as well.

What determines what is written to the log is determined by the
individual middlewares, and will be fleshed out in future commits.