Implement MiddlewareStorages into Middleware
Changes the Middleware and Middlewares to make use of the new
MiddlewareStorages.  All of the middleware plugins now call out to
save_report, sending it a filename, a shortform proc, and a longform
proc.  These arguments are then sent to each of the configured
MiddlewareStorage instances and they choose which of those arguments to
write to the storage backends.

For instance, the File MiddlewareStorage will make use of the filename
and the longform proc of each middleware and write their data to each of
their respective files, where as the Log MiddlewareStorage will only use
the shortform to aggregate that data to write it to the logfile.