Add "BrowserMode" configuration options
This will be a new feature for enabling the performance middleware
easily through the browser.  For now, it simply has configuration for
determining whether it is enabled and if it should be "always on"
(versus if it should be enabled via a request param).

Also, some fixes were needed for the `middleware_storage/log_spec`
because it overrides the default log location to keep the written log
file location nested under the `tmp` dir, but requires updating a bit of
hackery to do so.  Suggestions on doing this better welcome!