Fix reports for data without activesupport timers
Not all things that we will be reporting on will use the
ActiveSupportTimers middleware, or will use it and not have a result
because the data was gathered for a non-request based method.  This
means that the `ms` field might not always be available.

Since we are reporting after the fact, a small performance hit to find a
relevant column header will make this a much more versatile method, and
allow it to work went "ms" isn't available to be reported on.