Fix column_size_for for all requests
While we memoize `#column_size_for`, when it is first called for a
particular header, it only checks against data for a given request.
This can lead to an incorrect column size when moving on to a new
request.  This adds a fine grained memoization of the column size for
each result.