Fix averaging in reporter
Previously, the "avgs" column would just a sum of all of the data
collected for that particular metric, instead of an actually average.
It also wouldn't handle nil cases well when doing the `.inject`.

This fixes both and also reduces the code duplication a bit by creating
an `avg` function to make use of.