Fix saving ActiveSupport info when there is no data
This is not the most graceful solution, but it is the simplest for now.
The consequences of this is that there will be missing files for
requests using the file backend where this data doesn't exist, so that
is something to consider in the future.  For now, it should be unlikely.