Reparse URLs from filenames in reporter.rb
When using the `request_id` (needs a better name) to display above a
report, convert the %'s back to /'s so it looks like a URL again.

This change is also worth noting as a reminder that this reporter.rb
stuff was written when the `File` MiddlewareStorage was the only
available backend for the data.  This will need an interface for getting
data from other backends, so this change is mostly temporary.