Save all SQL params as strings
Some SQL parameters logged through ActiveSupport::Notifications are
actually full on classes, like ActiveSupport::TimeWithZone.  When
serialized to yaml, instead of these being in their "String" form, they
will be serialized as a ruby object.

This is a problem for anything digesting the yaml that doesn't have
these particular classes loaded.  To prevent this, everything becomes a
string when recording parameters for each call.