Prevent duplicate ActiveRecordQueries::Logger instances
In the future, we might be instantiating more than one instance of the
middleware to use it for something like a:

  ManageIQPerformance.profile {  do_stuff!  }

But because of how ActiveRecordQueries queries sets up a
ActiveSupport::Notifications.subscribe, duplicate subscription handlers
will write to the Thread variable for each handler that has been
initialized.

To prevent this, the private API of the
ActiveSupport::Notifications.notifier (Fanout) is queried against to
find the `listeners_for` the specific even that we are about to add a
notifier for.  If it our logger class isn't present, the extra will be
added, otherwise it will assume one is initialized already.