Adds ManageIQPeformance::profile
This is to be used for profiling something in line in irb, a script, or
a something like a Rails runner.

Uses the passed in block as the `app` for the middleware, and sets up a
dummy `env` that will make sure the middleware is triggered, and there
is a "name" for any file that is generated (if the file storage backend
is being used).  A name can also be passed in if the user wishes to name
the "request" for the run.

https://www.pivotaltracker.com/story/show/137479755