Fix missing instantiation.active_record subscriptions
When fixing duplicate ActiveSupport::Notifications subscriptions in
commit 623b2ab :

https://github.com/ManageIQ/manageiq-performance/commit/623b2ab

The `instantiation.active_record` subscription was removed entirely by
mistake because the check on existing subscriptions was checking against
the specific subscription, 'sql.active_record', instead of for the event
which was currently being set.

This resulted in the `rows` not being tallied when profiling code.