Generate timer info without a request
The ActiveSupportTimers middleware component requires a request to go
through if we wish to record data, but the data returned is actually not
that complex.  For things that just require some simple timers (like
::profile) , we can reorganize things slightly to allow this to capture
simple timer metrics without a request as well.

Also configures `ManageIQPerformance.profile` to make use of this new
timer.