GemBase64 lib for generating base64 .gem string
This lib is meant to be used for converting the result of the
manageiq-performance.gem to a base64 encoded string that can be used in
a installation script for an appliance.

This effectively avoids file IO by using a StringIO, and using a lot of
internals of the rubygems libs used for gem packaging.  The methods in
the specs are their to build the gem in a similar fashion and confirm
that the results are the same, but they when up being similar to the
methods they are testing...

The most crucial test going on here is the last one, which is testing
`gem_as_base64_string` and is being used to confirm building the gem and
encoding it is done properly, and can be decoded and unpacked to serve
back the same files.