Add rake task `generate_install_script`
This task is in charge of handling the automating of generating the
appliance installation script which allows manageiq-performance to be
installed on an appliance without bundler needing to be run.

More tests could be included around different versions of bundler and
handling the overwrite process.  For now, this gets something automated
and out there for future improvements to be made.

TODO:

* Optional arguments/ENV variables for changing the install directories
  and such on an appliance
* Testing portions of the script that are being generated (even though
  they will be just strings as far as the script generation process is
  concerned
* "appraisals" on different versions of bundler for the above
* Update travis to test on different versions of ruby supported versions
  MIQ is running on