Handle empty data[request_id][count_header] keys
As we add more middlewares, the types of keys being made available for
the headers is going to change with the middlewares that are activated
at the time.

While this is sort of a cover up for now, eventually we will have to
consider a method to determine the available headers for each
middleware being reported on.