Fix stackprof not being stopped in tests
When the ManageIQPerformance::profile method was added, some tests
confirming the middleware methods were executed in the correct order,
were copied from the original code testing the middleware `call` did the
same thing.

Unfortunately, the stubs were inplace to watch either the `start` or
`stop` methods that were being tested, but not the opposite ones that
weren't being tested.  This means that if stackprof was one of the
enabled middleware in the tests, and the program ended in a way that the
start was left un-stopped (I am assuming), then stackprof could "barf"
at the end of the test run, and cause an non-zero exit status do to a
segfault.

This unfortunately was a non-deterministic test failure, and even
running the same seed wouldn't necessarily reproduce the issue.  That
said, when running the test suite before this fix, you could run in to
the segfault within 25 runs of the suite.  Now, it doesn't seem to
happen after hundreds.