Add specs for Reporter#collect_data
Also fixes a precision issue when using pure floats when adding up the
elapsed times from the query data.  Since floats aren't accurate, we are
typecasting them to BigDecimal, and then converting them back again when
the timings have been added up.

The final result might be skewed slightly in the conversion, but it will
avoid the floating point issues.

A note about the files:

The data here is meant to be mostly stub data, but should be as close to
what would be pulled in a real situation.  Tried to reuse as much of it
as possible where possible and match up the print data seed results with
the yaml files.  There is a little bit of symlinking that was done to
reduce redundancy.  Only so much was possible though when trying to
catch all of the corner cases with this data.