Add include_stackprof task and update template
Updates the appliance_installation_script tasks to include a task called
"include_stackprof".  This task is meant to be used right before
generate_install_script like so:

    $ rake include_stackprof generate_install_script

This adds an few instance variables that will then be evaluated in the
template and added to the resulting output of the script.  If this isn't
run, then it is assumed that stackprof is not added, and that will not
be added to the generated script.

Because we are using instances variables in this task, it does mean that
the task HAS to be run with the generator task.  This shouldn't be a
huge issue, but something of note.