Remove X-CSRF-Token from full_request_headers
This never made sense because we were grabbing it from the homepage,
which isn't the same form that you would probably need this for when
authenticated in the app.  Thus, this is not useful.

What might be necessary in the future is to find the corresponding GET
page for a POST request to allow that POST to function.