Allow csrf_token to not exist
In https://github.com/ManageIQ/manageiq-ui-classic/issues/451 we removed
the CSRF token in our login page, which breaks the requestor code when
trying to log in and the csrf token is no where to be found.

By using the block form, we can return the proper match in the CSRF
token regexp if it exists, but still return `nil` if it does not, which
will be fine it the header is blank (which is why a `.to_s` was added to
the caller).

Instead of removing the csrf token lookup on login, we are keeping this
in place to allow this to be used on older versions of the MIQ
application, otherwise we would need conditional logic for different
versions of the application.