Add rake task support to profile command
Supports loading a rake task and executing it with it wrapped within the
profile code.

Loads the code in the same script instead of shelling out for the rake
task, which makes it more of an accurate profile, but might not fully
reflect the rake task environment.  More testing needed.