Allow for applying a temporary global config
Allowing the ManageIQPerformance.config to be changed for a period of
time, and allows context specific config changes to happen, but not
require a long term change with the config.

Useful with the `.profile` method so you can define config settings
changes within the context of the profile block, but then reset them
when you are finished.