Make `eq` consistent in config specs
DOWN WITH PARENTHESIS!!! (I like the way this looks way better...)

Also changed a few `match_array` to `eq` in there since `eq` doesn't
compare object_id and works as expected, but also because `match_array`
is order independent, and we needed the arrays to order dependent in
most of these checks.

The was a incorrectly passing test in a previous commit where the
`match_array` was causing a false positive  pass.