Adds memory storage
This storage simply takes the existing data from the profile runs, and
saves them into an ordered array as a tagged hash.  This storage is in
memory only and will not persist after the process has ended.

Also, added a way to access the last capture data via a top level
interface ManageIQPerformance.last_run.  The interface is only defined
if the user is using this middleware storage, otherwise it isn't
defined.