Adds :in_memory option to .profile
Allows for quickly accessing the memory middleware storage when using
the `.profile` method.

Current caveat is that the .profile method only works for 1 run, so
every time .profile is called, a new memory storage is created.

Also, in the tests, the FileUtils.rm_r was changed to an `rm_rf` because
when using the memory storage in these tests, a default_dir was not
created, so that call would cause specs fail.