Use Timecop to prevent non-deterministic failures
When testing the GemBase64 class, we compare the build of the gem
through normal means (building a gem and writing it to a file), and how
it is used in the GemBase64 class, which is saved to a StringIO object
in memory.

When the gem is tar'd up, it ends up (I assume) writing some timestamps
to the file, or some MD5SUMs (again, I assume that uses a timestamp in
the algorithm), which can cause differences in the output of the .gem
file and the StringIO outputs on occasion.

By using Timecop, this has completely prevented the issue from showing
up, unlike in master where this seems to break the tests every 5 runs or
so.