Refactor config_changes in .profile
To reduce the complexity of the `.profile` method, remove the optional
nature of using `with_config`, and always use it.  This removes the need
for needing an if around setting the config_changes variable.

As of note, there is still a need to have `config_changes` always be a
hash so that the `merge` can work in the next line when using the
`:in_memory` shortcut.