Add MiqQueueTimers middleware
This middleware is slightly different from other middlewares.

To start, it is only really meant to be included by the codebase, and
not by the user config (see the overwritten `deliver` method).

Also, it monkey patches the MiqQueue to inject a
`ManageIQPerformance.profile` call with every `deliver` call.  Part of
this monkey patch also includes the specialized MiqQueueTimers
middleware into the ManageIQPerformance middleware stack, which acts
similar to the active_support_timers middleware plugin.

Finally, this will also remove active_support_timers if that is
configured, since that will provide information that is not needed when
benchmarking the queue.