Update headers for miqperf benchmark --api
The default headers needed for calling to the API need to be requesting
json, and not html.  The current CONSTANTS wouldn't work for that, so
they needed to be renamed.

As such, all of the headers got new names, and the old names are now
deprecated, but still exist with there old values.  Probably not
necessary, but done for posterity.