Handle api auth/session in miqperf benchmark --api
Auth in the API doesn't require scraping for the csrf_token like it does
in the UI, but does require setting a different header value for the
auth, both when requesting the token for "login", and when making the
benchmarked request with the token known.

Also, 'json' is required now for the API, since the token is returned in
the response body and must be parsed.  This is dynamically loaded as
needed.