Fix/Tweak optparse requires in cli commands
In most of the miqperf commands, optparse was not being required (for
reasons I can't determine) in almost all of the command files.  This
fixes that, but also defers loading it until the option_parser command
has been run as well, which can make the default `help` command a bit
faster.