Refactor by adding .profile_env_for private method
This method takes a large chunk of the work being done in
`ManageIQPerformance.profile`, and does it in a private method.

A little more code (all of a `.detect` call) is now required to filter
out the callstack properly, but it feels a bit safer than it was
previously, so I don't mind it too much.

This is not a high traffic method either, so it doesn't really need a
performance optimization, and the method definition is marked as
private, which should shy away users from trying to use it from other
places.

The ROOT constant was also added to easily filter out the
ManageIQPerformance callstack lines in the `.detect`.  Kept this defined
in the `middleware.rb` file just because this is the only class using it
at the moment, but if something like the `CLI` or something finds a use
for it, I might push it up to the `lib/manageiq_performance.rb` file.

* * *

Finally, the main point of this was originally to remove some of the
more clunky tests around `ManageIQPerformance.profile` for testing this,
but I found that having both was nice for making sure I was keeping to
spec.  Might remove them in the future, but for now am just going to
keep the other ones I have.