Add spec/support/active_record_helper.rb
This is a helper context for rspec for dealing with spinning up a fake
schema to use for testing features of the active_record_queries
middleware.

This requires the `sqlite` gem to function, and on whatever
context/describe block the `:with_active_record_schema` is applied to,
it will set some before and after filters for all specs in the block to
spin up and down the schema properly.  Features include:

- Queries and new will be handled by the user.
- Includes two classes to start:  Book and Author
- Memory based DB, so nothing is written to the file system (fast)