Add ActiveRecordQueries.clear_memoized_logger_vars
Convenience method for clearing out instance variables that are set on
the ActiveRecordQueries::Logger instance.  Since only one of these can
exist at a time per event (though, usually the same instance for all of
them), this means it won't be re-initialized when a call to `.profile`
is used (for example).

This makes it so that in the future, we have a way of clearing out
variables we want to change between those runs (currently not hooked up
though).