Add build_c_ext_gem:debug[gem] task
This is used to boot the docker-machine instance with the container
running inside with the same setup that would be used with
`:build_c_ext_gem`.  This is helpful when debugging the issues with a
particular c-extension gem, and allow a terminal to be spun up in the
same environment as the automated build to provide a sandboxed debugging
environment.

The caveat with this is only some steps for this can be automated, and
the rest needs to be manually entered by the user to finish the setup.
This is "semi-automated" by the fact that the necessary commands needing
to be run are printed into the terminal prior to returning input back to
the user.  From there it is a copy paste task.