Add extension_dir fix
Unsure if this makes sense to keep, as I was using it as an experiment
when trying to build `rbtrace` with this workflow.  This did not work
because `rbtrace` was not properly standardized to work with
`rake-compiler`, and the extra hacks beyond this made it not worth
while.

That said, doesn't seem like it "should" break anything, so adding it
here for now.