Fix ruby loc in manageiq/ruby
There was a recent change in ManageIQ/container-ruby#4 that change the
ruby-install invocation to install ruby in the Dockerfile to be
`--system`.  This made it so that the config was not populated and all
`rake-compiler` tasks would fail.

This does cause a bit of a change to how the script generates the
version number and arch, but the output is the same.