Fix stackprof ext.lib_dir
Stackprof expects it's `stackprof.so` file in the `lib/stackporf/` dir,
but the default for `rake-compiler-dock` is to put it into `lib/`.

This fix allows not only for `stackprof` to properly configure specifics,
but exposes a way in the future for other c extension gems to make
changes to the Rakefile's `Rake::ExtensionTask` as well.