[requestor.rb] Fix #ignore_ssl_cert?
For whatever reason, when I coded this up, I didn't create a
`ignore_cert?` method that would match up with the instance variable. So
this library never worked for anything that was using port 443
(localhost requests worked fine though).

Was simpler to avoid adding the extra method and just used the instance
variable itself, so the fix was that.