[README] Add some badges
Updates the README.md to include badges that are currently setup for
this gem.  Ones left out (to be added later):

- Gem Version (not released on rubygems yet, so still a git gem)
- Code Coverage (don't have code coverage setup in the test suite yet)
- Dependency Status (invalid without this being on gitlab)
- Hakiri/Security (don't think the account is setup for this repo yet)

Also, updates `gitter` to a room that is currently valid (don't think we
need a special room for this repo, since the performance room doesn't
get enough traffic as it is).