[gem_base64_spec.rb] Fix for Gem::Package#build
In newer versions of rubygems, a small section of `Gem::Package#build`
relies on the `@gem`'s `@path` variable to be present and work with
`File.basename`.  Since this is not there, it fails and in previous
versions of it would only need to work with the `@spec` to get this
value.

Since we are doing a bit of fudging to get tihs to work, this change
just implements the parts of `Gem::Package#build` that we for our
purposes instead of trying more hacks to that method to do what we need
it to.