[gem_base64_spec.rb] Refactor untar
Instead of returning the contents of a single file based off a pattern,
return the entire contents of the gem package as a Hash.

This allows the similar functionality it was previously used for
(extracting the file contents of a single file in a gem), but also
diffing the entire contents in a more user/developer friendly fashion
(versus diffing the raw contents).