[gem_base64_spec.rb] Debug ::gem_as_tar_io spec
Adds a second expectation to the "::gem_as_tar_io" spec that uses the
`untar` method to build out a Hash of the contents of the gem package
and the reference gem, and compares those hashes.  This can be used to
determine why the specs are failing much easier than comparing the raw
compressed/encoded form.