[gemspec] Fix gem_base64_spec.rb failures
Because of OS sorting randomness, the order which these files are
returned from the glob can be in different orders, causing the
metadata.gz within the gem package to be different between different
builds of a given gem package.

By sorting in Ruby after we get the results, we can eliminate that
variable from being an issue.